+++ packages/elysia/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/elysia
 
+## 0.1.64
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.6
+
+## 0.1.63
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.5
+
 ## 0.1.62
 
 ### Patch Changes