+++ packages/koa/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/koa
 
+## 0.1.47
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.6
+
+## 0.1.46
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.5
+
 ## 0.1.45
 
 ### Patch Changes